skills/id-photo-maker/requirements.txt
Pillow>=10.0.0
rembg>=2.0.0
